requirements.txt
gunicorn==19.9.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
coveralls==1.5.1
Django==2.1.4
django-cors-middleware==1.3.1
django-extensions==2.1.4
djangorestframework==3.9.0
psycopg2==2.7.6.1
docopt==0.6.2
idna==2.8
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
six==1.12.0
whitenoise==4.1.2
urllib3==1.24.1
flake8==3.6.0
